langserve-fastapi-server/requirements.txt
langchain
langchain_openai
langchain_core
python-dotenv
langserve
fastapi
uvicorn
